[INLONG-9614][Agent] Change folder name
