+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/sinks/ProxySink.java
@@ -24,9 +24,9 @@ import org.apache.inlong.agent.constant.CommonConstants;
 import org.apache.inlong.agent.core.task.MemoryManager;
 import org.apache.inlong.agent.core.task.OffsetManager;
 import org.apache.inlong.agent.message.EndMessage;
-import org.apache.inlong.agent.message.filecollect.OffsetAckInfo;
-import org.apache.inlong.agent.message.filecollect.ProxyMessage;
-import org.apache.inlong.agent.message.filecollect.SenderMessage;
+import org.apache.inlong.agent.message.file.OffsetAckInfo;
+import org.apache.inlong.agent.message.file.ProxyMessage;
+import org.apache.inlong.agent.message.file.SenderMessage;
 import org.apache.inlong.agent.plugin.Message;
 import org.apache.inlong.agent.plugin.MessageFilter;
 import org.apache.inlong.agent.plugin.sinks.filecollect.SenderManager;
