+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/task/file/WatchEntity.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.agent.plugin.task.filecollect;
+package org.apache.inlong.agent.plugin.task.file;
 
 import org.apache.inlong.agent.plugin.utils.file.DateUtils;
 import org.apache.inlong.agent.plugin.utils.file.FilePathUtil;
