+++ inlong-agent/agent-plugins/src/test/java/org/apache/inlong/agent/plugin/task/TestLogfileCollectTask.java
@@ -23,7 +23,7 @@ import org.apache.inlong.agent.constant.TaskConstants;
 import org.apache.inlong.agent.core.task.TaskManager;
 import org.apache.inlong.agent.db.Db;
 import org.apache.inlong.agent.plugin.AgentBaseTestsHelper;
-import org.apache.inlong.agent.plugin.task.filecollect.LogFileCollectTask;
+import org.apache.inlong.agent.plugin.task.file.LogFileTask;
 import org.apache.inlong.common.enums.TaskStateEnum;
 
 import com.google.gson.Gson;
@@ -50,13 +50,13 @@ import java.util.concurrent.TimeUnit;
 import static org.awaitility.Awaitility.await;
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest(LogFileCollectTask.class)
+@PrepareForTest(LogFileTask.class)
 @PowerMockIgnore({"javax.management.*"})
 public class TestLogfileCollectTask {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(TestLogfileCollectTask.class);
     private static final ClassLoader LOADER = TestLogfileCollectTask.class.getClassLoader();
-    private static LogFileCollectTask task;
+    private static LogFileTask task;
     private static AgentBaseTestsHelper helper;
     private static final Gson GSON = new Gson();
     private static TaskManager manager;
@@ -90,7 +90,7 @@ public class TestLogfileCollectTask {
             taskProfile.setLong(TaskConstants.TASK_START_TIME, start);
             taskProfile.setLong(TaskConstants.TASK_END_TIME, end);
             manager = new TaskManager();
-            task = PowerMockito.spy(new LogFileCollectTask());
+            task = PowerMockito.spy(new LogFileTask());
             PowerMockito.doAnswer(invocation -> {
                 fileName = invocation.getArgument(0);
                 dataTime = invocation.getArgument(1);
